pyproject.toml
[project]
name = "testagent"
version = "0.1"
description = "testagent agent"
readme = "README.md"
dependencies = [
  "bedrock-agentcore>=1.6.2",
  "mcp-proxy-for-aws>=1.3.0",
  "strands-agents>=1.35.0",
]
requires-python = ">=3.13"

[project.scripts]
testagent = "testagent.agent:main"
